codegen.generate_fixture -> fixture.create_fixture
